requirements.txt
ollama
ruff
pyyaml
